We now import PVC data too so we label based on VA etiology.
We grab lesion numbers from GLG instead of our own records. When we don't have reliable footage for dwell time data, we use GLG data.

We manually calculate ablation time for patient 9 based on hybrid our records and GLG records.

We calculate energy on time

For patients that didn't complete the study such as 1-011 and 1-021, instead of dropping them we set their data to all NAs such that when we do summary statistics they're still reflected in the sample size but also show up as "missing"